backend/api/views.py
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework import status
import numpy as np
import pandas as pd
from datetime import datetime, timedelta

class StockPredictionView(APIView):
    def post(self, request):
        # Placeholder for your ML model prediction
        # You'll integrate your trained model here
        return Response({
            "prediction": {
                "price": 180.45,
                "confidence": 0.85,
                "trend": "upward"
            }
        }, status=status.HTTP_200_OK)

class MarketingPredictionView(APIView):
    def post(self, request):
        # Placeholder for your marketing ML model
        # You'll integrate your trained model here
        return Response({
            "prediction": {
                "recommended_channels": ["social_media", "email"],
                "expected_roi": 2.87,
                "target_audience": "tech-savvy professionals"
            }
        }, status=status.HTTP_200_OK)