backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
python-dotenv==1.0.1
gunicorn==21.2.0